Interactions example change proposal
I have changed solution to interactions issue. I know that it requires knowledge of bind, but I think if one is reading this part of is series is familiar with it regardless whether one has read previous books or not. This example shows more robust and cleaner JavaScript code which is in a spirit of this series.